backend/requirements.txt
uvicorn
fastapi[standard]
opensearch-py
websockets
python-dotenv